补充
git-svn-id: https://tinn-pc:8443/svn/Common/Paway.Forms@374 c1914438-a9fc-4b41-8dc5-b03b8268c31d

+++ Paway.Helper/Common/RowComparer.cs
@@ -65,11 +65,11 @@ namespace Paway.Helper
                         s2 += arr2[j];
                         j++;
                     }
-                    if (int.Parse(s1) > int.Parse(s2))
+                    if (long.Parse(s1) > long.Parse(s2))
                     {
                         return 1;
                     }
-                    if (int.Parse(s1) < int.Parse(s2))
+                    if (long.Parse(s1) < long.Parse(s2))
                     {
                         return -1;
                     }
